removing shared folder
