+++ README.md
@@ -20,7 +20,7 @@ $ npm install -g veloce-sfdx-v3
 $ sfdx COMMAND
 running command...
 $ sfdx (-v|--version|version)
-veloce-sfdx-v3/0.0.5 darwin-x64 node-v16.15.1
+veloce-sfdx-v3/0.0.6 darwin-x64 node-v16.15.1
 $ sfdx --help [COMMAND]
 USAGE
   $ sfdx COMMAND
@@ -72,7 +72,7 @@ EXAMPLE
   ./source/templates
 ```
 
-_See code: [src/commands/veloce/data/pull.ts](https://github.com/veloceapps/veloce-sfdx-v3/blob/v0.0.5/src/commands/veloce/data/pull.ts)_
+_See code: [src/commands/veloce/data/pull.ts](https://github.com/veloceapps/veloce-sfdx-v3/blob/v0.0.6/src/commands/veloce/data/pull.ts)_
 
 ## `sfdx veloce:data:push [-p <string>] [-v <string>] [-u <string>] [--apiversion <string>] [--json] [--loglevel trace|debug|info|warn|error|fatal|TRACE|DEBUG|INFO|WARN|ERROR|FATAL]`
 
@@ -105,7 +105,7 @@ EXAMPLE
   --sourcepath ./source/templates
 ```
 
-_See code: [src/commands/veloce/data/push.ts](https://github.com/veloceapps/veloce-sfdx-v3/blob/v0.0.5/src/commands/veloce/data/push.ts)_
+_See code: [src/commands/veloce/data/push.ts](https://github.com/veloceapps/veloce-sfdx-v3/blob/v0.0.6/src/commands/veloce/data/push.ts)_
 
 ## `sfdx veloce:debug:list [-v <string>] [-u <string>] [--apiversion <string>] [--json] [--loglevel trace|debug|info|warn|error|fatal|TRACE|DEBUG|INFO|WARN|ERROR|FATAL]`
 
@@ -135,7 +135,7 @@ EXAMPLE
   sfdx veloce:debug:list --targetusername myOrg@example.com --targetdevhubusername devhub@org.com
 ```
 
-_See code: [src/commands/veloce/debug/list.ts](https://github.com/veloceapps/veloce-sfdx-v3/blob/v0.0.5/src/commands/veloce/debug/list.ts)_
+_See code: [src/commands/veloce/debug/list.ts](https://github.com/veloceapps/veloce-sfdx-v3/blob/v0.0.6/src/commands/veloce/debug/list.ts)_
 
 ## `sfdx veloce:debug:logs [-v <string>] [-u <string>] [--apiversion <string>] [--json] [--loglevel trace|debug|info|warn|error|fatal|TRACE|DEBUG|INFO|WARN|ERROR|FATAL]`
 
@@ -165,7 +165,7 @@ EXAMPLE
   sfdx veloce:debug:logs --targetusername myOrg@example.com --targetdevhubusername devhub@org.com
 ```
 
-_See code: [src/commands/veloce/debug/logs.ts](https://github.com/veloceapps/veloce-sfdx-v3/blob/v0.0.5/src/commands/veloce/debug/logs.ts)_
+_See code: [src/commands/veloce/debug/logs.ts](https://github.com/veloceapps/veloce-sfdx-v3/blob/v0.0.6/src/commands/veloce/debug/logs.ts)_
 
 ## `sfdx veloce:debug:push [-v <string>] [-u <string>] [--apiversion <string>] [--json] [--loglevel trace|debug|info|warn|error|fatal|TRACE|DEBUG|INFO|WARN|ERROR|FATAL]`
 
@@ -195,7 +195,7 @@ EXAMPLE
   sfdx veloce:debug:push --targetusername myOrg@example.com --targetdevhubusername devhub@org.com
 ```
 
-_See code: [src/commands/veloce/debug/push.ts](https://github.com/veloceapps/veloce-sfdx-v3/blob/v0.0.5/src/commands/veloce/debug/push.ts)_
+_See code: [src/commands/veloce/debug/push.ts](https://github.com/veloceapps/veloce-sfdx-v3/blob/v0.0.6/src/commands/veloce/debug/push.ts)_
 
 ## `sfdx veloce:debug:start [-u <string>] [--apiversion <string>] [--json] [--loglevel trace|debug|info|warn|error|fatal|TRACE|DEBUG|INFO|WARN|ERROR|FATAL]`
 
@@ -222,7 +222,7 @@ EXAMPLE
   sfdx veloce:debug:start --targetusername myOrg@example.com --targetdevhubusername devhub@org.com
 ```
 
-_See code: [src/commands/veloce/debug/start.ts](https://github.com/veloceapps/veloce-sfdx-v3/blob/v0.0.5/src/commands/veloce/debug/start.ts)_
+_See code: [src/commands/veloce/debug/start.ts](https://github.com/veloceapps/veloce-sfdx-v3/blob/v0.0.6/src/commands/veloce/debug/start.ts)_
 
 ## `sfdx veloce:debug:stop [-v <string>] [-u <string>] [--apiversion <string>] [--json] [--loglevel trace|debug|info|warn|error|fatal|TRACE|DEBUG|INFO|WARN|ERROR|FATAL]`
 
@@ -252,7 +252,7 @@ EXAMPLE
   sfdx veloce:debug:stop --targetusername myOrg@example.com --targetdevhubusername devhub@org.com
 ```
 
-_See code: [src/commands/veloce/debug/stop.ts](https://github.com/veloceapps/veloce-sfdx-v3/blob/v0.0.5/src/commands/veloce/debug/stop.ts)_
+_See code: [src/commands/veloce/debug/stop.ts](https://github.com/veloceapps/veloce-sfdx-v3/blob/v0.0.6/src/commands/veloce/debug/stop.ts)_
 
 ## `sfdx veloce:debug:watch [-v <string>] [-u <string>] [--apiversion <string>] [--json] [--loglevel trace|debug|info|warn|error|fatal|TRACE|DEBUG|INFO|WARN|ERROR|FATAL]`
 
@@ -282,7 +282,7 @@ EXAMPLE
   sfdx veloce:debug:watch --targetusername myOrg@example.com --targetdevhubusername devhub@org.com
 ```
 
-_See code: [src/commands/veloce/debug/watch.ts](https://github.com/veloceapps/veloce-sfdx-v3/blob/v0.0.5/src/commands/veloce/debug/watch.ts)_
+_See code: [src/commands/veloce/debug/watch.ts](https://github.com/veloceapps/veloce-sfdx-v3/blob/v0.0.6/src/commands/veloce/debug/watch.ts)_
 
 ## `sfdx veloce:fixref [-d] [-u <string>] [--apiversion <string>] [--json] [--loglevel trace|debug|info|warn|error|fatal|TRACE|DEBUG|INFO|WARN|ERROR|FATAL]`
 
@@ -313,7 +313,7 @@ EXAMPLES
   sfdx veloce:fixref --targetusername myOrg@example.com --targetdevhubusername devhub@org.com
 ```
 
-_See code: [src/commands/veloce/fixref.ts](https://github.com/veloceapps/veloce-sfdx-v3/blob/v0.0.5/src/commands/veloce/fixref.ts)_
+_See code: [src/commands/veloce/fixref.ts](https://github.com/veloceapps/veloce-sfdx-v3/blob/v0.0.6/src/commands/veloce/fixref.ts)_
 
 ## `sfdx veloce:login -p <string> -a <string> -u <string> -r <string> [-s <string>] [--json] [--loglevel trace|debug|info|warn|error|fatal|TRACE|DEBUG|INFO|WARN|ERROR|FATAL]`
 
@@ -346,7 +346,7 @@ EXAMPLE
   $ sfdx veloce:login -u username -p ./PASSWORDFILE -a alias01
 ```
 
-_See code: [src/commands/veloce/login.ts](https://github.com/veloceapps/veloce-sfdx-v3/blob/v0.0.5/src/commands/veloce/login.ts)_
+_See code: [src/commands/veloce/login.ts](https://github.com/veloceapps/veloce-sfdx-v3/blob/v0.0.6/src/commands/veloce/login.ts)_
 
 ## `sfdx veloce:source:pull [-m <string>] [-p <string>] [-v <string>] [-u <string>] [--apiversion <string>] [--json] [--loglevel trace|debug|info|warn|error|fatal|TRACE|DEBUG|INFO|WARN|ERROR|FATAL]`
 
@@ -382,7 +382,7 @@ EXAMPLE
   --sourcepath ./source/pmls
 ```
 
-_See code: [src/commands/veloce/source/pull.ts](https://github.com/veloceapps/veloce-sfdx-v3/blob/v0.0.5/src/commands/veloce/source/pull.ts)_
+_See code: [src/commands/veloce/source/pull.ts](https://github.com/veloceapps/veloce-sfdx-v3/blob/v0.0.6/src/commands/veloce/source/pull.ts)_
 
 ## `sfdx veloce:source:push [-m <string>] [-p <string>] [-v <string>] [-u <string>] [--apiversion <string>] [--json] [--loglevel trace|debug|info|warn|error|fatal|TRACE|DEBUG|INFO|WARN|ERROR|FATAL]`
 
@@ -418,5 +418,5 @@ EXAMPLE
   --sourcepath ./source/pmls
 ```
 
-_See code: [src/commands/veloce/source/push.ts](https://github.com/veloceapps/veloce-sfdx-v3/blob/v0.0.5/src/commands/veloce/source/push.ts)_
+_See code: [src/commands/veloce/source/push.ts](https://github.com/veloceapps/veloce-sfdx-v3/blob/v0.0.6/src/commands/veloce/source/push.ts)_
 <!-- commandsstop -->
