+++ src/utils/productModel.utils.ts
